restructured glcomp cleared more warnings

+++ lib/glcomp/glutils.c
@@ -30,9 +30,9 @@ int GetFixedOGLPos(int x, int y, float kts, GLfloat * X, GLfloat * Y,
     GLfloat winX, winY;
     GLdouble posX, posY, posZ;
 
-	glColor4f(0,0,0,0.001);
+	glColor4f((GLfloat)0,(GLfloat)0,(GLfloat)0,(GLfloat)0.001);
 	glBegin(GL_POINTS);
-    glVertex3f(10.00, 10.00, 1.00);
+    glVertex3f((GLfloat)10.00, (GLfloat)10.00, (GLfloat)1.00);
     glEnd();
 	
 	glGetDoublev(GL_MODELVIEW_MATRIX, modelview);
@@ -74,7 +74,7 @@ int GetOGLPosRef(int x, int y, float *X, float *Y, float *Z)
     glGetIntegerv(GL_VIEWPORT, viewport);
 
     //draw a point  to a not important location to get window coordinates
-	glColor4f(0,0,0,0.001);
+	glColor4f((GLfloat)0,(GLfloat)0,(GLfloat)0,(GLfloat)0.001);
 
 	glBegin(GL_POINTS);
     glVertex3f(10.00, 10.00, 0.00);
@@ -121,7 +121,7 @@ float GetOGLDistance(int l)
     glGetIntegerv(GL_VIEWPORT, viewport);
 
     //draw a point  to a not important location to get window coordinates
-	glColor4f(0,0,0,0.001);
+	glColor4f((GLfloat)0,(GLfloat)0,(GLfloat)0,(GLfloat)0.001);
 
 	glBegin(GL_POINTS);
     glVertex3f(10.00, 10.00, 1.00);
@@ -161,7 +161,7 @@ int GetFixedOGLPoslocal(int x, int y, float kts, GLfloat * X, GLfloat * Y,
     glGetDoublev(GL_PROJECTION_MATRIX, projection);
     glGetIntegerv(GL_VIEWPORT, viewport);
 
-	glColor4f(0,0,0,0.001);
+	glColor4f((GLfloat)0,(GLfloat)0,(GLfloat)0,(GLfloat)0.001);
 	glBegin(GL_POINTS);
     glVertex3f(10.00, 10.00, 0.00);
     glEnd();
